+++ user-profile/src/main/java/com/microservice/erp/domain/repositories/IChangeEmailVerificationCodeRepository.java
@@ -3,5 +3,7 @@ package com.microservice.erp.domain.repositories;
 import com.microservice.erp.domain.entities.ChangeEmailVerificationCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IChangeEmailVerificationCodeRepository extends JpaRepository<ChangeEmailVerificationCode, Long> {
- }
+import java.math.BigInteger;
+
+public interface IChangeEmailVerificationCodeRepository extends JpaRepository<ChangeEmailVerificationCode, BigInteger> {
+}
