+++ user-profile/src/main/java/com/microservice/erp/domain/repositories/IChangeMobileNoSmsOtpRepository.java
@@ -3,5 +3,7 @@ package com.microservice.erp.domain.repositories;
 import com.microservice.erp.domain.entities.ChangeMobileNoSmsOtp;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IChangeMobileNoSmsOtpRepository extends JpaRepository<ChangeMobileNoSmsOtp, Long> {
- }
+import java.math.BigInteger;
+
+public interface IChangeMobileNoSmsOtpRepository extends JpaRepository<ChangeMobileNoSmsOtp, BigInteger> {
+}
